requirements.txt
streamlit>=1.30.0
langchain>=0.2.0
langchain-groq>=0.1.0
langgraph>=0.1.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
arxiv>=2.1.0
PyMuPDF>=1.23.0
pyvis>=0.3.2
reportlab>=4.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
langchain-core>=0.2.0
